Move all FIELDS for api calls into models (except messages, which doesn't have a model yet)
