+++ src/providers/uploadHelpers/resizeImage.js
